backend/utils/jwt_utils.py
import jwt
import os
from datetime import datetime, timedelta

SECRET = os.getenv("JWT_SECRET", "supersecretkey")

def generate_token(user_id):
    payload = {
        "user_id": str(user_id),
        "exp": datetime.utcnow() + timedelta(days=1)
    }
    return jwt.encode(payload, SECRET, algorithm="HS256")

def decode_token(token):
    try:
        return jwt.decode(token, SECRET, algorithms=["HS256"])
    except jwt.ExpiredSignatureError:
        return None
